Add temporary fix for CodeReloader in test env
`Mix.Task.reenable` cannot be run in the test environment as it causes
subsequent test runs to break on defined module errors for the main
application module. The result requires a complete `rm -rf _build` to fix.

As a temporary fix until a better solution is found, instead of running
the full reenable/recompile mix task for the test env, `IEx.Helpers.r` is
applied to all modules that respond to `phoenix_recompile?`.
This allows files such as Views to be dynamically recompiled for TDD of
views/templates and satisfies the only code reloading required for the test env
today. The dev env code reloading remains unchanged. Close #317