Initial router changes
I was a bit baffled as to why the router Phoenix generated at first was
so different from the one depicted in the guide. While it did not throw
me off long due to a Ruby/Rails background, someone completely new to
routers might be more puzzled than I was so this is my attempt at
clarifying that.