View and layout functions raise AlreadySentError
It's easy to think that the layout and view is set when calling `plug
:action` in a controller so people (myself included) put their plugs
below `plug :action` and then wonder why `put_layout` and friends don't
work as expected. In reality the view and layout are set when you call
`use Phoenix.Controller` so user's plugs should go above `plug :action`
if you're working with the view or layout.

This PR makes it so that when you call `put_layout`, `put_new_layout`,
`put_view`, `put_new_view` and `put_layout_formats` on a conn that was
already sent, you get a friendly error instead of wondering why the
wrong view/layout was rendered.