Remove "clearly"/"obviously" from Ecto models guide
Using language like "clearly", "obviously", and
"basically" is a common habit and is not intended
to be negative, but it can be offputting when
found in guidance as the concepts described are
frequently neither clear nor obvious for newcomers.

It's also redundant, as if it was obvious we wouldn't
be putting it in the documentation.