Switch validation example in Ecto models guide
This change removes the split of "first name" and
"last name" with just a "name" field, and adds a
"bio" field to use as an example for length
validation.

There's two main reasons for this change:
1. The idea of a neat first name and last name
   is not as universal as we might think, and the
   benefit gained from having separate fields is
   small against the cost of confusion and annoyance.
2. There's no practical reason for putting a length
   limit on a name field, especially one shorter
   than the database field it's going into.  It's
   not a great experience for those of us with
   long surnames.

These sorts of cultural assumptions are not
implemented with malice in mind, just habit,
but they provide little benefit to us as programmers
or to our products, and often only serve to make
those from other cultures feel unwelcome.

We should wherever possible avoid spreading these
antipatterns to other developers.

There's some more discussion on name fields here:
https://www.gov.uk/service-manual/user-centred-design/resources/patterns/names.html