Require pubsub adapters to define a dispatch table
Prior to this commit, every PubSub operation had to
be serialized through a server, often only to delegate
to a local server.

This commit solves this issue by requiring PubSub
adapters to define a dispatch table with the same name
as the server and reading directly from this table
when a dispatch is necessary.