Fix latin abbreviations in Phoenix.js comments
Previously, the comments read "ie", which is short for "id est", meaning, "that is". This implies that what follows is what you're really trying to say. For example, on line 211, ie meant that the event was always going to be `"phx_join"`.
Now, the comments read "e.g.", which is short for "exempli gratia", meaning "for example". That same example now means that event is going to be a string from the list of events, like `"phx_join"`