Make mix generators supply string keys for controller params
Controllers are expected to receive parameters with string keys, but
the generators generate tests with atom keys. At best this means the
user has to re-write them, at worst it can be misleading about what
the standard is.

This commit makes `Mix.Phoenix.params` return a map with string
keys, and adds a method `Mix.Phoenix.expected_attr` that returns the
same thing with atom keys.

It then amends the generated tests to rename `@valid_attrs` and
`@invalid_attrs` to `@*_params`, and adds `@expected_attrs`. The tests
themselves are then amended to send `@*_params` to the controller, and
test against `@expected_attrs`.