make token `max_age` check use seconds
Prior to this commit, we had a disconnect between the documented units
(seconds) and the actual units (milliseconds) that were used for
validating the maximum age of a token. This meant that the suggestion of
a 2 week maximum age as 1209600 seconds was, in reality, checked as
milliseconds, resulting in a default maximum age of a little over 20
minutes. This commit updates the time used in both the token and the
check to be seconds, consistent with the documentation.