Fixes to the testing introduction guide
* Remove the "==> ranch (compile)" and  "Generated hello_phoenix app"
lines and the three dots from the first test run output to avoid
confusion with the four dots for the running tests. Now the output
matches the output you'd get after having compiled before, like the rest
of the test output in the guide.
* Add `test/views/layout_view_test.exs` to the `tree test` output
* Remove the "Create the database..."-comment in test/test_helper.exs,
as it's been removed in
https://github.com/phoenixframework/phoenix/commit/fcdc47a5ae19c00fe00ccb83c717c3c95af8685e.
* Add `end` tags to the `test/views/error_view_test.exs` code blocks
* Include the full output for `$ mix test
test/views/error_view_test.exs:12`
* Add a `$` before `mix test --exclude error_view_case`
* Include the `import Phoenix.View`-line and match indentation to the
generated file in `test/views/error_view_test.exs`
* Remove double "value" in "We can also specify a value and run only
tests with that (value)."