Use receive timeout hooks to replace after clauses
Closes #1295 #1265 #1241

`push` now accepts an optional timeout which will
trigger `receive("timeout", ...)` blocks. This replaces
`after` clauses as it requires explicit timeout handling
with a sane 15s default. This change also only buffers
unsent messages between disconnects during the timeout window.

Leaving a channel when a join failed is now fixed, but an optimistic
"leave" event is still pushed to the server in order to solve races
where the joined timed out for the client, but succeeded on the server.