Don't explain how to change the template root directory
While interesting to know, explaining how to change the template root
directory is out of the scope of the templates guide, as changing the
template root is something you should probably only do when you're more
familiar with the framework. Also, putting it at the start of the guide
like this is confusing, as it's not completely clear if the reader
should do anything with the template root directory or not. It might be
a good idea to move this elsewhere.