Clarify how to use the Phoenix compiler
I recently used Phoenix.View in a library which required the Phoenix
compiler. It raised this helpful error, but I wasn't sure how to add it
to compilers. When I did find `Mix.compilers` I did this:

    Mix.compilers ++ [:phoenix]

Which didn't work. It turns out that `phoenix` needs to be first. This
commit adds an example to make things a bit easier.