feat(Phoenix.Router.Helpers): better errors on path helpers
Instead of simply returning a FunctionClauseError when passing an atom
for an undefined action to a helper, the error now prompts the user
with a list of valid atoms.

The `ArgumentError` is raised with the following message:

    No route helper clause for foo_path/2 defined for action :skip.
    The following foo_path actions are defined under your router:

     * :file
     * :show

A slightly different error is displayed if the helper and action are
valid, but the arity is incorrect.