to_string for put_resp_header
Was following the guide and the example for "put_headers" produced the error: no function clause matching in Plug.Conn.put_resp_header/3.

I figured this was because the key of the map was an atom, and put_resp_header expects a binary. So I converted to atom to_string and it worked.