Remove subscribe/1 and unsubscribe/1 from Channel
Fastlaning external topics is a misfeature as the
client is only listening to their active
channel topic and cannot handle fastlaned external
messages. The correct approach is what was favored
previously; using subscribing via the endpoint API
and picking the broadcast up in handle_info and
finally forwarding to the client with whatever
transforms necessary.