Add workaround for `exrm` bug
`exrm` has a recent incompatibility with `mix` that is detailed in [this issue](https://github.com/bitwalker/exrm/issues/326). This update to the guide provides a workaround to that glitch by running `mix compile` before `mix release`.

Note that once the `exrm` issue is resolved, this step can be removed. I had a more verbose explanation of the error in the guide, but figured that the guide should be as clear and simple as possible for new users, so I worked it in as though it's a normal step.