Fix prod.exs to include defaults for exrm builds
It's a shame that to get phoenix just running we need to dig in docs.
Align config with documentation: http://www.phoenixframework.org/docs/advanced-deployment

It's not compulsory but will save hours to a newcomers who will deploy by creating release. Exrm releases are de facto standard way to deploy.