Adds Presence.get_by_key with default
Sometimes we want just a single piece of information out of the Presence
tracker. In order to facilitate that we want to be able to grab from the
tracker with a key and not get the entire map. I mapped this to the same
behavior as `Map.get` so it would have the ability to give a default if
one doesn't exist. By default `Map.get` returns `nil` if something isn't
in the map. Since we are expecting map to come back from Presence I
decided to make this default into an empty map.

Amos King @adkron <amos@binarynoggin.com>