Keep table names simple
Namespacing the table name with the context prefix is causing
a lot of confusion in regards management of references and
associations. Furthermore, it arguably couples the code with
the database. In the previous approach, would you rename the
table if you rename the prefix?

The only potential gain of the previous behaviour is in case
someone wants to have different users tables per context but
such cases can be handled as needed by using the --table option.

For this reason, we are rolling back the table names to simply
be the plural name given to the generator, as in earlier Phoenix
versions.