Update instructions for token authentication in comments (#2334)
The instructions currently refer to the old way to passing passing
parameters on `socket.connect`. If folled, this results in a
warning message in the JS console that "passing params to connect is
deprecated. Instead pass :params to the Socket constructor", see
[`phoenix.js`](https://github.com/phoenixframework/phoenix/blob/master/assets/js/phoenix.js#L533).

This confused me as I was reading through this, so I've created an
update that points to the correct location for passing the token.