Promote short-lived Phoenix tokens by default
The token is mostly used only for authenticating the connection,
so there is no reason for long periods. If the user disconencts
and the token expired, then the user needs to be instructed to
reload the page - this is already the recommended practice, for
example in case the user is deleted, unauthorized and so forth.

We now also warn if there is no max_age, in order to promote
better defaults.