Remove the GroupLeaderProxy and re-use the IO handling logic in Proxy to handle both devices.
When output capture is started a Task is spawned to temporarily replace the target PID’s group leader, and watch for IO requests using a receive loop.

This will then call the now shared `handle_io` function, which will send a message to the Proxy process if there is any output. This avoids both the Proxy and the Task loop having to keep their own state, instead the Proxy handles keeping track of the output and forwarding it to the observers.