Change Proxy to avoid issues with other users replacing `standard_error`.
A test was failing due to the use of `capture_io` which also replaces the `standard_error`, resulting in an error when the Proxy was used as it expects to be running under that name.

This change has the code reload `Server` store the pid of the proxy and have calls to the proxy `GenServer` use the PID directly instead of the `standard_error` name, to avoid such errors.