Update Cowboy 2 handler and websocket
The Cowboy 2 handler functions in the same way as the cowboy handler,
gathering the sockets and transports and creating routes for them.
Instead of falling back to the `Plug.Adapters.Cowboy.Handler` module, it
falls back to `Plug.Adapters.Cowboy2.Handler`.

For Cowboy 2, the websocket pid when connecting to the transport. This
is passed to connect, requiring a change in arity for the
Transport.connect function. It defaults to self() for backwards
compatability.

The integration tests have been updated to toggle the handler being used
depending on the version of Cowboy defined in the lockfile.