Update the docs for bypass_through to be more explicit.
I found it hard to understand that invoking the route/endpoint was related to the use of `get("/")` and ran into a `** (ArgumentError) flash not fetched, call fetch_flash/2` error on my tests related to the fact I had a flash dependency in my test but was not invoking the router.  That led me to write a post in the elixir forums. https://elixirforum.com/t/bypass-through-3-not-working/10070/2
 
It was later made clear that I needed to invoke the router via `get("/")`. 

I made this pr in hopes to guide people to this same solution when searching for `flash not fetched, call fetch_flash/2` since I was unable to resolve my issue by searching alone and felt the docs missed explicitly calling out why the use of `get("/")` was needed.