Add security headers (#2692)
X-Download-Options

The X-Download-Options HTTP header has only one option:
X-Download-Options: noopen. This is for Internet Explorer from
version 8 on to instruct the browser not to open a download
directly in the browser, but instead to provide only the ‘Save’
option. The user has to first save it and then open it in an
application. The reason for this is that when HTML files
download from the application, the browser will render it directly
inline. That means that the file renders as part of the application
and has direct access to the security context of it, meaning it may
run phishing attacks or maybe access critical domain cookies.

More info at:

http://msdn.microsoft.com/en-us/library/ie/jj542450(v=vs.85).aspx

X-Permitted-Cross-Domain-Policies

The X-Permitted-Cross-Domain-Policies HTTP header restricts Adobe
Flash player’s access to data. The Adobe products PDF and Flash use
a same-origin concept that sometimes allow cross-domain access.
Flash on a.com will be able to access data on b.com if there is a
Cross Domain Policy File at b.com/crossdomain.xml that allows data
to be accessed from a.com (or from all sources).

More info at:

https://www.adobe.com/devnet/adobe-media-server/articles/cross-domain-xml-for-streaming.html