Rework the introduction and organization of the Channels guide
Having gotten lost in some of these abstractions myself, I tried to make
this introduction as concrete and clear as possible. The examples and
new diagram at the top should help readers quickly grasp what channels
can be used for and the flexibility and scalability they offer.

Beyond that, I added a few clarifications and examples, and reordered
the parts so that PubSub is last, since it's the least visible to an
application developer.