Remove mix files from guides
Previously, the guides were a mix project as they were maintained separately.
Now that the guides are built as part of the Phoenix docs, there is no need for
this separation. The guides now contains all of the markdown files at the top
level. All duplicate files such as `LICENCE` and `README.md` have been removed
as they are covered by the Phoenix project as a whole.