Add changelog link to new template
The phoenix version string that was previously used in the dependency
string has been moved to a module attribute so it can be easily reused.
The `Version` module is used so that both "1.4.0" and "1.4" formats can
easily be generated. The latter format is used when linking to the
changelog in the template.

A test has also been added to ensure the correct version is specified in
mix.exs based on the version of the installer.