Cache the pubsub server value
The lookup has been moved to the supervisor which is used when computing the
value to be cached. Caching is required because this value is looked up every
time there is a pubsub broadcast from the endpoint. The pubsub name is
fetched on warmup so that the value is available in the cache from when
the supervisor is started.