Only warn on config if it is not set and unchanged after init
Previously, if the config was modified in the init/2 callback it would still
need to be specified in the config. This would require setting the config to an
empty list to suppress the warning:

    config :my_app, MyApp.Endpoint, []

This warning is now only applied of the init callback returns `{:ok, config}`
without any modifications AND the config is not set via the application config.