Allow runtime configuration of PubSub (#3071)
* Allow runtime configuration of PubSub

Previously PubSub was required at compile time. This commit allows configuration
at runtime by looking up the PubSub module for the endpoint from the config
instead of compiling it into the module. The compile time errors have been move
to runtime and will trigger when attempting to start the pubsub children.

* Cache the pubsub server value

The lookup has been moved to the supervisor which is used when computing the
value to be cached. Caching is required because this value is looked up every
time there is a pubsub broadcast from the endpoint. The pubsub name is
fetched on warmup so that the value is available in the cache from when
the supervisor is started.