Update plug to 1.6.4 for mix-cowboy1.lock
The dependencies have been manually updated in the lockfile to prevent
cowboy and friends being upgraded when fetching a version of plug
greater than that already in the lock file. In this case, the lockfile
was previously set to plug `1.6.0`, but mix.exs was at `1.6.2`. When
fetching the dependencies for running the tests, plug was being upgraded
to `1.6.3` or greater, triggering the cowboy upgrade.