Fix building docs with babel 7
documentation.js required version 6 of babel, but phoenix has migrated to
version 7. To fix this, the dependency has been updated to
v9.0.0-alpha.0 which supports the latest babel.

Although this is an alpha dependency, it is specific to building the
docs, so as long as they build correctly, I don't see a huge issue with
it, it is preferable to downgrading babel.