Use FunctionClauseError for Phoenix.ActionClauseError
Using `FunctionClauseError` results in showing the available clauses
when displaying the error message. This also renders nicely in the plug
exceptions page.

Instead of using `plug_status` in the exception, the `Plug.Exception`
protocol is explicitly defined for the `ActionClauseError` to allow easy
transitioning between `ActionClauseError` and `FunctionClauseError`.`

This closes #2657