Allow custom paths to be provided to :websocket and :longpoll (#3166)
* Allow custom paths to be provided to :websocket and :longpoll

Previously, it was not possible to use a socket with a path that differed from
the key name. For example `:websocket` would always end with `/websocket` and
`:longpoll` would always end with `/longpoll`.

This commit allows a `:path` to be passed into the keyword options, the path
will be used instead of the default if available.