Allow variables to be used when defining socket paths
This commit allows sockets (websocket/longpoll) to use variables in the path.
For example, a socket can be defined with:

    socket "/ws/:user_id", MyApp.UserSocket,
      websocket: [path: "/project/:project_id"]

The params will be available in the connect map:

    def connect(%{"user_id" => user_id, "project_id" => project_id}), do: ...

The `__handler__/2` function now hands off the a `do_handler` which is
responsible for modifying the passed `conn` to add the `path_params` and
`params`. `__handler__/2` is now called with the entire `conn` instead of just
the `path_info` and expects a `conn` to be available in the returned tuple.

As this is only supported in Cowboy 2, the `test_helper.exs` has been
updated to ignore `:cowboy2` tests when testing with `plug_cowboy` 1.x.