Fix errors_on when using array field changeset
## Overview
`errors_on/1` raises an error when it receives a changeset with an array
field which was `cast`-ed with an invalid type.

## Details
When validating the type of an array field fails, `cast/4` it will put an error on the changeset, such as:
```elixir
[my_list_field: {"is invalid", [type: {:array, :string}, validation: :cast]}]
```

While processing this error, `errors_on/1` will try to `to_string/1` the
`{:array, :string}` tuple, which produces the error:
```
** (Protocol.UndefinedError) protocol String.Chars not implemented for {:array, :string}. This protocol is implemented for: Atom, BitString, Date, DateTime, Decimal, Float, Integer, List, NaiveDateTime, Postgrex.Copy, Postgrex.Query, Time, URI, Version, Version.Requirement
```

## Solution
The implemented solution is to skip trying to interpolate error message embeds
when the value to be embedded is a tuple.

## Test Case
I didn't write a test for this, as I'm not sure how/where to do so for
installer templates. I tested locally with a setup like:
```elixir
defmodule MyAppWeb.Drzewo do
  use Ecto.Schema
  import Ecto.Changeset

  embedded_schema do
    field :kolory, {:array, :string}
  end

  def changeset(attrs) do
    cast(%Drzewo{}, attrs, [:kolory])
  end
end

defmodule MyAppWeb.DrzewoTest do
  use MyApp.DataCase

  alias MyAppWeb.Drzewo

  test "errors_on raises an error when passed an invalid type for the array field" do
    drzewo_params = %{kolory: 1}

    changeset = Drzewo.changeset(drzewo_params)
    changeset.valid?

    errors_on(changeset)
  end
end
```