Drop special convention for context locations
The original intent was to highlight the context file
as the interface to the code contained within its
directory. While a worthwhile pursuit, the convention
has caused more confusion than it has helped. At the
same time, we can achieve our interface goals with
existing guides to drive best practices.