Add phoenix_route private
This commit adds an easy way in Phoenix to retrieve the path of the route
against which a request was matched. Previously, if you for example defined a
`get "/users/:id"` route and you receive a `GET /users/1` request, you had no
easy solution to retrieve that the route was `/users/:id`. You’d only have
access to `conn.request_path` which would be something like `/users/1` in such
case.

With this commit, you can call `Phoenix.Router.match_path(conn)` to retrieve
the route's path.

Accessing the route can be really useful in the following cases:

* `logging` - logging your requests by adding a route metadata in order to
              filter quite easily by route.
* `metrics` - tagging your request metrics with the route so that you could
              for example check the performance of your app per route.
* `rate_limiting` - creating a route-based rate limiting.

This feature is similar to a feature implemented in Plug
(https://github.com/elixir-plug/plug/pull/630).