Add option to install Postgres.app
I received a `DBConnection.ConnectionError` after following the instructions in Installation, the Up and Running Guide, and running `mix ecto.create`. I resolved the problem by installing Postgres.app and starting the Postgres server before running Ecto or Phoenix. If the error is not a bug, this solution could be presented as an option for Mac users to get up and running.