Fix rejoining channels after explicit reconnect
Previously when a socket which had channels open was explicitly
reconnected (`socket.disconnect(() => socket.connect())` then the
existing channels were not rejoined.

This was because `onConnClose` was not called. This was due to a timing
issue in `teardown` function. This function called close on connection
and also cleaned up onclose listener on the connection object. The close
however happened in the next event loop by which time the onclose
listener was already removed.

The channels stayed in `joined` state even though socket was
disconnected. On reconnect `joined` channels were not rejoined.

Fixes #3458