Update default socket serializer to use V2
Since v1.4, the [socket documentation states that `Phoenix.Socket.V2.JSONSerializer' is used by default](https://github.com/phoenixframework/phoenix/blob/v1.4/lib/phoenix/socket.ex#L89-L92).

However, that is not the case because when the `vsn` parameter is excluded, it gets set to `"1.0.0"` making the default V1.JSONSerializer.

This changes that to be `"2.0.0"` when `vsn` is not explicitly provided in the connect params in order to match documentation and use `V2.JSONSerializer` by default.