Silence context generation warning on --no-context
When a context exists and a user issues `mix phx.gen.live --no-context`,
there is a warning about generating into an existing context, even
though the user is not using the context. This commit prevents that
warning from displaying.

**Note:** The code being generated by the generator is not changing. It
is what the user would expect (files without a context).

The test in this commit, without the production code applied, outputs
the following:

```
 ** (RuntimeError) no shell process input given for yes?/1
 stacktrace:
   (mix 1.10.4) lib/mix/shell/process.ex:148: Mix.Shell.Process.yes?/1
   (phoenix 1.6.0-dev) lib/mix/tasks/phx.gen.context.ex:279: Mix.Tasks.Phx.Gen.Context.prompt_for_code_injection/1
   (phoenix 1.6.0-dev) lib/mix/tasks/phx.gen.live.ex:98: Mix.Tasks.Phx.Gen.Live.run/1
   test/mix/tasks/phx.gen.live_test.exs:288: anonymous fn/0 in Mix.Tasks.Phx.Gen.LiveTest."test with --no-context does not emit warning when context exists"/1
   (elixir 1.10.4) lib/file.ex:1544: File.cd!/2
   installer/test/mix_helper.exs:36: MixHelper.in_tmp_project/2
   test/mix/tasks/phx.gen.live_test.exs:282: (test)

```