Mention 'npm install --prefix assets' in more instructions
It's a little easier than cd:ing into the dir, running it and then going back.

Some guides mention an alternative on Windows, so I did the same here.

Also changing some '--prefix ./assets' to '--prefix assets' for consistency.

Did not change

> Phoenix developers typically only need to run `npm install` inside the assets directory.

in guides/directory_structure.md because it seemed like too much detail for that doc.

Did not change references in error output in lib/phoenix/endpoint/watcher.ex because it
seemed more of a one-off, and there a portable but less convenient command seems better.